+++ friendtracker/Shared/Utils/Theme/ColorSchemeManager.swift
@@ -1,0 +1,16 @@
+import SwiftUI
+
+class ColorSchemeManager: ObservableObject {
+    @AppStorage("isDarkMode") var isDarkMode = false {
+        didSet {
+            colorScheme = isDarkMode ? .dark : .light
+        }
+    }
+    @Published var colorScheme: ColorScheme = .light
+    
+    static let shared = ColorSchemeManager()
+    
+    private init() {
+        colorScheme = isDarkMode ? .dark : .light
+    }
+}